+++ paddle/phi/kernels/xpu/activation_grad_kernel.cc
@@ -209,7 +209,7 @@ struct XPULogGradFunctor : public funcs::BaseActivationFunctor<T> {
 
 template <typename T>
 struct XPULeakyReluGradFunctor : public funcs::BaseActivationFunctor<T> {
-  float alpha;
+  double alpha;
   typename funcs::BaseActivationFunctor<T>::AttrPair GetAttrs() {
     return {{"alpha", &alpha}};
   }
